테스트 코드 작성 Mockito, 단위 테스트 공부 필요성 테스트 코드 트랜젝션 롤백 동작 시 PK의 AUTO INCREMENT 값은 롤백되지 않음

+++ src/main/java/toyproject/juniorforum/controller/BoardController.java
@@ -32,14 +32,14 @@ public class BoardController {
     }
 
     @GetMapping("/{boardId}")
-    public String read(@PathVariable Long boardId, Model model) {
+    public String read(@PathVariable int boardId, Model model) {
         BoardVO board = boardService.read(boardId);
         model.addAttribute("board", board);
         return "board/read";
     }
 
     @GetMapping("/{boardId}/update")
-    public String updateForm(@PathVariable Long boardId, Model model) {
+    public String updateForm(@PathVariable int boardId, Model model) {
         BoardVO boardVO = boardService.read(boardId);
         BoardDTO board = boardVO.convertToDTO();
         model.addAttribute("board", board);
